convert MenuItem.js to typescript
